- DTO (Fix) - DAO (Fix) - Service & Repo Fixes & Improvements - Controller & WebController Redesign - Converter Implementation - (Mongo & Jpa) (Repo & Service) works - etc

+++ src/main/java/com/univ/db/repo/ItemRepository.java
@@ -15,5 +15,5 @@ import java.util.Optional;
  */
 @Repository
 public interface ItemRepository extends MongoRepository<Item, String> {
-    Optional<Item> findByCatalogCode(Integer code);
+    Optional<Item> findByCatalogCode(Long code);
 }
